refactor: Refactored source to accommodate unit testing
