+++ src/program/index.js
@@ -1,0 +1,3 @@
+import program from "./program";
+
+export default program;
