+++ src/tasks/index.js
@@ -1,0 +1,5 @@
+import tasks from "./tasks";
+
+export { createTasks } from "./tasks";
+
+export default tasks;
